Resolves "Mapping - Add Apps to all TrafficLights" (#26)
It is now possible to add multiple apps distributed by weights to traffic lights. If the parameters"weight" and "tlGroupId" in the traffic light mapping, then the default value of the weight is 1. The mapping schema was adapted, so that weights in the traffic light mapping can be bigger than 1

* Fixed mapping of traffic light apps
* adapted schema so that all items are validated
* adjusted test to fixed json schema
* added test for traffic light in spawning framework init

Signed-off-by: Franziska Vogt <franziska.vogt@fokus.fraunhofer.de>