Send vehicle name to remove form the visualizer separately
WebsocketVisualizerServer sends only a limited amount of VehicleUpdates interactions in order to avoid slowing down the simulation for visualization purposes. This implementation drops VehicleUpdates including the list of removed vehicles, resulting in vehicle markers not be removed from the map canvas. This is now fixed by collecting all vehicles to be removed and send them separately to the web visualizer in the browser.

Signed-off-by: Karl Schrab <karl.schrab@fokus.fraunhofer.de>